SYSFW-NAVSS-ResAssg: j721e: Partition GPIO Interrupt routers
All outputs of the GPIO IRs (both main and wakeup instances)
are not routable to all CPUs. Few lines are connected to a set
of CPUs and few are not. This adds a constraint where a certain
range of resources can only be assigned to a set of HOST_IDs

To achieve this, split the range into multiple ranges and highlight
the cell where it is legal to assign the resources.

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>