dfu-boot: Fix UART boot mode selection for am64xx-evm
am64xx-evm does not work correctly when switched to uart boot mode.
The board is successfully put in the uart boot mode but then
the uniflash script fails to send bootloader to the ROM.

When the bootmode is switched from A53 u-boot prompt, it works correctly.
So add a change_bootmode_slow function which does this.
Use this function only for am64xx-evm uart boot mode.

Signed-off-by: Naresh A <nareshk@ti.com>
Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>